move fetchCustomPrompt to consumer file
